change names of files and classes

+++ assignments/assignment-2/const.py
@@ -4,7 +4,7 @@ USER_TABLE_INSERT = "id, has_labels"
 
 ACTIVITY_TABLE_NAME = "activity"
 ACTIVITY_TABLE_SCHEMA = """
-    id BIGINT NOT NULL,
+    id INT NOT NULL,
     user_id CHAR(3) NOT NULL,
     transportation_mode VARCHAR(10) NULL,
     start_date_time DATETIME(0) NOT NULL,
@@ -17,7 +17,7 @@ ACTIVITY_TABLE_INSERT = "id, user_id, transportation_mode, start_date_time, end_
 TRACK_POINT_TABLE_NAME = "track_point"
 TRACK_POINT_TABLE_SCHEMA = """
     id INT AUTO_INCREMENT NOT NULL,
-    activity_id BIGINT NOT NULL,
+    activity_id INT NOT NULL,
     lat DOUBLE(9, 6) NOT NULL,
     lon DOUBLE(9, 6) NOT NULL,
     altitude INT NULL,
